backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.28.85
python-jose[cryptography]==3.3.0
reportlab==4.0.4
pandas==2.1.1